Initial commit
Artificial intelligence is not just a smart system: it's a smart system capable of self-learning. In this project, you will create a simple interactive game where the computer will try to guess the animal that the person has in mind with the help of yes or no questions. During the game, the computer will extend its knowledge base by learning new facts about animals and using this information in the next game.